Move transformer loading code to a separate file
